% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageing_comparison_plot.R
\name{ageing_comparison}
\alias{ageing_comparison}
\title{Plot comparison of double age readings}
\usage{
ageing_comparison(xvec, yvec, scale.pts = 2, col.pts = grey(0.1, alpha =
  0.5), col.hist = rgb(0, 0, 0.5, alpha = 0.7), counts = TRUE,
  maxage = NULL, hist = TRUE, hist.frac = 0.1, xlab = "Age reader A",
  ylab = "Age reader B", title = NULL)
}
\arguments{
\item{xvec}{vector of values from reader A}

\item{yvec}{vector of values from reader B}

\item{col.pts}{color for points}

\item{col.hist}{color for histograms}

\item{counts}{include text within each bubble showing count of values?}

\item{maxage}{maximum age to include in the plot (doesn't yet work well)}

\item{hist}{include a histogram along each axis?}

\item{hist.frac}{maximum value of histograms as fraction of maxage}

\item{xlab}{label for xvec}

\item{ylab}{label for yvec}

\item{title}{Optional title to add at top of plot}
}
\description{
Plot with circles proportional to how many double readings
fell in each pair of coordinates
}
\author{
Ian G. Taylor
}
