% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeingError.R
\docType{package}
\name{AgeingError-package}
\alias{AgeingError}
\alias{AgeingError-package}
\title{Ageing error software}
\description{
R package to support the Punt et al. (2008) ageing error model
}
\references{
Punt, A.E., Smith, D.C., KrusicGolub, K., and Robertson, S. 2008.
Quantifying age-reading error for use in fisheries stock assessments,
with application to species in Australias southern and eastern scalefish and shark fishery.
Canadian Journal of Fisheries and Aquatic Sciences 65: 1991-2005.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pfmc-assessments.github.io/AgeingError/}
  \item Report bugs at \url{https://github.com/pfmc-assessments/AgeingError/issues}
}

}
\author{
\strong{Maintainer}: Ian G. Taylor \email{ian.taylor@noaa.gov} (\href{https://orcid.org/0000-0002-4232-5669}{ORCID})

Authors:
\itemize{
  \item Ian G. Taylor \email{ian.taylor@noaa.gov} (\href{https://orcid.org/0000-0002-4232-5669}{ORCID})
  \item Andre E. Punt (\href{https://orcid.org/0000-0001-8489-2488}{ORCID})
  \item Kelli F. Johnson (\href{https://orcid.org/0000-0001-9563-1937}{ORCID})
  \item Paul Burch (\href{https://orcid.org/0000-0002-9853-462X}{ORCID})
}

Other contributors:
\itemize{
  \item James T. Thorson (\href{https://orcid.org/0000-0001-7415-1010}{ORCID}) [contributor, copyright holder]
  \item Ian J. Stewart [contributor]
  \item Melissa A. Haltuch (\href{https://orcid.org/0000-0003-2821-1267}{ORCID}) [contributor]
}

}
\keyword{internal}
