% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMx.R
\name{cMx}
\alias{cMx}
\title{Make a column matrix}
\usage{
cMx(Input)
}
\arguments{
\item{Input}{input to be converted to a matrix}
}
\description{
\subsection{The function is currently defined as}{

function (Input)
{
as.matrix(Input)
}
}
}
\author{
James T. Thorson
}
