% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{CreateData}
\alias{CreateData}
\title{Deprecated function renamed to load_data()}
\usage{
CreateData(...)
}
\arguments{
\item{...}{Any arguments associated with the deprecated function}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
CreateData() has been renamed as \code{\link[=load_data]{load_data()}} to better clarify its purpose
}
\seealso{
\code{\link[=load_data]{load_data()}}
}
\author{
Ian G. Taylor
}
