% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSpecs.R
\name{CreateSpecs}
\alias{CreateSpecs}
\title{Read the ageing error specifications}
\usage{
CreateSpecs(SpecsFile = "data.spc", DataSpecs, verbose = FALSE)
}
\arguments{
\item{SpecsFile}{Filename for input specifications.}

\item{DataSpecs}{The output from CreateData()}

\item{verbose}{Return messages to the console (TRUE/FALSE)}
}
\description{
Read the ageing error specifications
}
\author{
Andre E. Punt
}
