% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine-.R
\name{determine_n_sets}
\alias{determine_n_sets}
\title{Determine the number of data sets in a data file}
\usage{
determine_n_sets(file)
}
\arguments{
\item{file}{A file path to a data file.}
}
\value{
An integer giving the number of data sets in the file.
}
\description{
Determine the number of data sets in a data file
}
\author{
Kelli F. Johnson
}
