% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoApplyAgeError.R
\name{DoApplyAgeError}
\alias{DoApplyAgeError}
\title{Run the ageing error optimization routine}
\usage{
DoApplyAgeError(
  Species = "AgeingError",
  DataSpecs,
  ModelSpecsInp,
  AprobWght = 1e-06,
  SlopeWght = 0.01,
  SaveDir = getwd(),
  verbose = FALSE
)
}
\arguments{
\item{Species}{A string that will be used to create file names. Typically,
users will use the common name for the species of interest, especially if
you are saving files from multiple species in a single directory. Though,
the default is \code{"AgeingError"}.}

\item{DataSpecs}{A data object returned from \code{\link[=load_data]{load_data()}}.}

\item{ModelSpecsInp}{A specification object returned from \code{\link[=load_specs]{load_specs()}}.}

\item{AprobWght, SlopeWght}{Numeric values passed to the model. The defaults
are 1e-06 and 0.01. Andre originally had these hard coded from his
workspace. TODO: decide if they should be passed in the specifications or
data files.}

\item{SaveDir}{A path, relative or absolute, to a directory where the
results will be saved. The directory need not exist currently as it will
be created dynamically.}

\item{verbose}{A logical specifying if messages should be printed. The
default is to \strong{NOT} print, i.e., \code{verbose = FALSE}.}
}
\description{
Run the ageing error optimization routine
}
\author{
Andre E. Punt
}
