% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load the formatted ageing error data}
\usage{
load_data(DataFile = "data.dat", NDataSet = 1, verbose = FALSE, EchoFile = "")
}
\arguments{
\item{DataFile}{Filename for input data}

\item{NDataSet}{Number of data sets within \code{DataFile}}

\item{verbose}{Return messages to the console (in addition to any output to
\code{EchoFile})}

\item{EchoFile}{A file path to a file that will be created or appended to if
it already exists to store information about your data inputs. The default
is \code{''}, which leads to output being printed to the screen rather than
saved in a file. An example of a user-defined input would be
\code{'EchoTMB.out'}.}
}
\description{
Load the formatted ageing error data
}
\author{
Andre E. Punt
}
