% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_specs.R
\name{load_specs}
\alias{load_specs}
\title{Read the ageing error specifications}
\usage{
load_specs(SpecsFile = "data.spc", DataSpecs, verbose = FALSE)
}
\arguments{
\item{SpecsFile}{Filename for input specifications.}

\item{DataSpecs}{The output from load_data()}

\item{verbose}{Return messages to the console (TRUE/FALSE)}
}
\description{
Read the ageing error specifications
}
\author{
Andre E. Punt
}
