% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimizer.R
\name{minimizer}
\alias{minimizer}
\title{Minimize the negative log likelihood}
\usage{
minimizer(
  model,
  method = c("optim", "nlmimb", "both"),
  lower,
  upper,
  verbose = FALSE
)
}
\arguments{
\item{model}{A model to be optimized.}

\item{method}{A string specifying the desired method to be used for the
optimization routine. The options are listed in the function call, where
the default is to use \code{"optim"}. Using both routines is an option, via
\code{"both"}, and will lead to first optimizing the model using \code{"nlminb"}
and then re-optimization of the model with \code{"optim"}. Note that when using
\code{\link[stats:optim]{stats::optim()}}, the \code{"L-BFGS-B"} method is used rather than the default
method of \code{"Nelder-Mead"}.}

\item{lower, upper}{Vectors of parameter bounds of the same length as the
number of parameters in the model.}

\item{verbose}{A logical specifying if messages should be printed. The
default is to \strong{NOT} print, i.e., \code{verbose = FALSE}.}
}
\description{
Minimize the negative log likelihood using \code{"nlmimb"} and/or \code{"optim"}.
}
\author{
Andre E. Punt
}
