% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOutputFn.R
\name{PlotOutputFn}
\alias{PlotOutputFn}
\title{Deprecated function replaced by plot_output()}
\usage{
PlotOutputFn(...)
}
\arguments{
\item{...}{Any arguments associated with the deprecated function}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
PlotOutputFn() has been replaced by \code{\link[=plot_output]{plot_output()}}
}
\seealso{
\code{\link[=plot_output]{plot_output()}}
}
\author{
James T. Thorson, Ian G. Taylor
}
