% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessResults.R
\name{ProcessResults}
\alias{ProcessResults}
\title{Process results of the ageing error estimation}
\usage{
ProcessResults(
  Species = "AgeingError",
  SaveDir = getwd(),
  CalcEff = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Species}{A string that will be used to create file names. Typically,
users will use the common name for the species of interest, especially if
you are saving files from multiple species in a single directory. Though,
the default is \code{"AgeingError"}.}

\item{SaveDir}{A path, relative or absolute, to a directory where the
results will be saved. The directory need not exist currently as it will
be created dynamically.}

\item{CalcEff}{Calculate effective sample sizes (TRUE/FALSE)}

\item{verbose}{A logical specifying if messages should be printed. The
default is to \strong{NOT} print, i.e., \code{verbose = FALSE}.}
}
\description{
Process results of the ageing error estimation
}
\author{
Andre E. Punt
}
