% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMx.R
\name{rMx}
\alias{rMx}
\title{Make a row matrix}
\usage{
rMx(Input)
}
\arguments{
\item{Input}{input to be converted into a row matrix}
}
\description{
\subsection{The function is currently defined as}{

function (Input)
{
if (is.vector(Input)) {
Output <- t(as.matrix(Input))
}
if (!is.vector(Input)) {
Output <- as.matrix(Input)
}
Output
}
}
}
\author{
James T. Thorson
}
