% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunFn.R
\name{RunFn}
\alias{RunFn}
\title{Deprecated function replaced by run()}
\usage{
RunFn(...)
}
\arguments{
\item{...}{Any arguments associated with the deprecated function}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
RunFn() has been replaced by \code{\link[=run]{run()}}
}
\seealso{
\code{\link[=run]{run()}}
}
\author{
James T. Thorson, Ian J. Stewart, Andre E. Punt, Ian G. Taylor
}
