% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFn.R
\name{SimulatorFn}
\alias{SimulatorFn}
\title{Simulate double reading data}
\usage{
SimulatorFn(Nreaders, M, SelexForm, ErrorParams, BiasParams, SelexParams,
  ReadsMat, RecCv = 0.6, RecAr1 = 0.8, Amax = 100)
}
\arguments{
\item{Nreaders}{The number of ageing readers}

\item{M}{True natural mortality}

\item{SelexForm}{Form of selectivity}

\item{ErrorParams}{Error type}

\item{BiasParams}{Bias type}

\item{SelexParams}{Selectivity parameters}

\item{ReadsMat}{Matrix describing number of reads per reader combination}

\item{RecCv}{CV of recruitment}

\item{RecAr1}{First-order autoregressive coefficient for recruitment}

\item{Amax}{True maximum age}
}
\value{
Returns a simulated double read matrix
}
\description{
A function to generate simulated double reading data with given properties
}
\references{
Punt, A.E., Smith, D.C., KrusicGolub, K., and Robertson, S. 2008.
Quantifying age-reading error for use in fisheries stock assessments,
with application to species in Australias southern and eastern scalefish
and shark fishery. Can. J. Fish. Aquat. Sci. 65: 1991-2005.
}
\author{
James T. Thorson
}
