% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_repeats.R
\name{tally_repeats}
\alias{tally_repeats}
\title{Tally repeated age reading combinations in a new count column}
\usage{
tally_repeats(dat)
}
\arguments{
\item{dat}{Input dataframe or tibble with columns for each reader and rows
for each specimen}
}
\description{
AgeingError requires a table of data with columns for each reader and
rows for each age reading combination, where a \code{count} column indicates
how many times the age reading combination is repeated. This function
tallies the repeated values and adds an initial \code{count} column with the
tally.
}
\author{
Ian G. Taylor
}
