% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_file.R
\name{write_data_file}
\alias{write_data_file}
\title{Write a data file for the AgeingError package}
\usage{
write_data_file(
  dat,
  dir = getwd(),
  file_name = "data.dat",
  minage = 0,
  maxage = NULL,
  refage = NULL,
  minusage = NULL,
  plusage = NULL
)
}
\arguments{
\item{dat}{Dataframe or tibble with columns for each reader and rows for
each age reading combination. This could either have a count column or
not. If not, \code{\link[=tally_repeats]{tally_repeats()}} will be called to add a count column.
Order your reader/lab columns such that similar
readers/labs are located next to one another because columns to the right
can mirror columns to their immediate left in terms of parameter
estimates.}

\item{dir}{Directory where the data file will be saved.}

\item{file_name}{Name of the data file.}

\item{minage}{An integer, specifying the minimum possible "true" age.}

\item{maxage}{An integer, specifying the maximum possible "true" age.}

\item{refage}{An arbitrarily chosen age from which "true" age-composition
fixed-effects are calculated as an offset. This has no effect on the
answer but could potentially effect estimation speed. By default this will
be set to the maxage / 4.}

\item{minusage}{The minimum age for which an age-specific age-composition is
estimated. Ages below \code{minusage} have "true" proportion-at-age
(\eqn{P_{a}}) estimated as
\deqn{P_a = P_{minusage}*exp^{(\beta*(minusage - a))}},
where beta is an estimated log-linear trend in the "true"
proportion-at-age. If \code{minusage} = \code{minage}, beta is not estimated.}

\item{plusage}{Identical to \code{minusage} except defining the age above with
age-specific age composition is not estimated.}
}
\value{
Invisibly returns the path to the data file (\code{file.path(dir, file_name)}).
}
\description{
The resulting file has the format required by the \code{\link[=load_data]{load_data()}} function.
In the future, this step could be bypassed by creating the list output
from \code{\link[=load_data]{load_data()}} directly. Only one data set is supported (NDataSet = 1).
This function is based on a subset of the RunFn() function in the older ADMB
version of this package.
}
\examples{
data_test <- data.frame(
  reader1 = c(7, 10, 7, 6, 6, 10, 7, 9, 8, 10, 10, 5, 6, 7, 9, 7, 7, 5, 8, 5),
  reader2 = c(8, 10, 7, 6, 6, 10, 7, 9, 8, 10, 10, 5, 6, 7, 9, 7, 7, NA, NA, NA),
  reader3 = c(7, 10, 7, 6, 6, 8, 7, 9, 8, 10, 10, 5, 6, 7, NA, NA, NA, 5, 8, 5)
)
data_file <- write_data_file(data_test, dir = tempdir(), file_name = "test.dat")

}
\seealso{
\code{\link[=write_files]{write_files()}}, \code{\link[=write_specs_file]{write_specs_file()}}, \code{\link[=load_data]{load_data()}}, \code{\link[=tally_repeats]{tally_repeats()}}
}
\author{
Ian G. Taylor, James T. Thorson, Ian J. Stewart, Andre E. Punt
}
