% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R
\name{write_files}
\alias{write_files}
\title{Write the input files (data and specifications) for the AgeingError package}
\usage{
write_files(
  dat,
  dir = getwd(),
  file_dat = "data.dat",
  file_specs = "data.spc",
  minage = 0,
  maxage = NULL,
  refage = NULL,
  minusage = NULL,
  plusage = NULL,
  biasopt = NULL,
  sigopt = NULL,
  knotages = NULL
)
}
\arguments{
\item{dat}{Dataframe or tibble with columns for each reader and rows for
each age reading combination. This could either have a count column or
not. If not, \code{\link[=tally_repeats]{tally_repeats()}} will be called to add a count column.
Order your reader/lab columns such that similar
readers/labs are located next to one another because columns to the right
can mirror columns to their immediate left in terms of parameter
estimates.}

\item{dir}{Directory where the data file will be saved.}

\item{minage}{An integer, specifying the minimum possible "true" age.}

\item{maxage}{An integer, specifying the maximum possible "true" age.}

\item{refage}{An arbitrarily chosen age from which "true" age-composition
fixed-effects are calculated as an offset. This has no effect on the
answer but could potentially effect estimation speed. By default this will
be set to the maxage / 4.}

\item{minusage}{The minimum age for which an age-specific age-composition is
estimated. Ages below \code{minusage} have "true" proportion-at-age
(\eqn{P_{a}}) estimated as
\deqn{P_a = P_{minusage}*exp^{(\beta*(minusage - a))}},
where beta is an estimated log-linear trend in the "true"
proportion-at-age. If \code{minusage} = \code{minage}, beta is not estimated.}

\item{plusage}{Identical to \code{minusage} except defining the age above with
age-specific age composition is not estimated.}

\item{biasopt}{A vector with one entry for each reader specifying the
type of bias specific to each reader. Positive values lead to estimated
parameters and negative values are used for shared parameters between
readers. Parameter sharing (mirroring) is common when there
is more than one reader in a lab working together to refine their methods
such that they have matching techniques.
If NULL is passed, the default is \code{rep(0, nreaders)} which specifies that
all readers are unbiased.

Possible entries include the following:
\describe{
\item{-\link{0-9}+}{
Mirror the bias of another reader, where the negative integer
corresponds to the column of the reader that is being mirrored
minus one, e.g., \code{-1} causes it to mirror reader 1. Only lower-numbered
readers can be mirrored.
}
\item{0}{
Unbiased, where at least one reader has to be unbiased.
}
\item{1}{
Constant coefficient of variation, i.e., a 1-parameter linear
relationship of bias with true age.
}
\item{2}{
Curvilinear, i.e., a 2-parameter Hollings-form relationship of bias
with true age.
}
}

An example entry for the situation where you have seven readers and you
assume that the first reader is unbiased, readers 2-7 have a curvilinear
bias, reader 3 shares parameters with reader 2, reader 5 shares parameters
with reader 4, and reader 7 shares parameters with reader 6 would look
like \code{c(0, 2, -2, 2, -4, 2, -6)}.}

\item{sigopt}{A vector with one entry for each reader.
Each entry specifies the functional
form of reading error as a function of true age. Positive values lead to
estimated parameters and negative values are used for shared parameters
between readers.
If NULL is passed, the default is \code{c(1, rep(-1, nreaders - 1))} which
specifies a constant CV in ageing error which is shared among all readers.

Possible entries include the following:
\describe{
\item{-\link{0-9}+}{
Mirror the standard deviation of another reader, where the negative
integer corresponds to the column of the reader that is being
mirrored minus one, e.g., \code{-1} causes it to mirror reader 1, for
which data is stored in the second column of \code{Data}. This number must
be lower than -1 times the current position in the vector.
}
\item{0}{
No error. But, there could be potential bias.
}
\item{1}{
Constant coefficient of variation, i.e., a 1-parameter linear
relationship of the standard deviation with the true age.
}
\item{2}{
Curvilinear standard deviation, i.e., a 3-parameter Hollings-form
relationship of standard deviation with true age.
}
\item{3}{
Curvilinear coefficient of variation, i.e., a 3-parameter
Hollings-form relationship of coefficient of variation with true age.
}
\item{5}{
Spline with estimated slope at beginning and end where the number of
parameters is 2 + number of knots.
}
\item{6}{
Linear interpolation with a first knot of 1 and a last knot of the
maximum age, i.e., \code{MaxAge}.
}
}}

\item{knotages}{Ages associated with each knot. This is a necessary input
for \code{sigopt = 5} or \code{sigopt = 6}. Not implemented in this function yet.}
}
\value{
Invisibly returns the path to the data file (\code{file.path(dir, file_name)}).
}
\description{
Write the input files (data and specifications) for the AgeingError package
}
\examples{
data_test <- data.frame(
  reader1 = c(7, 10, 7, 6, 6, 10, 7, 9, 8, 10, 10, 5, 6, 7, 9, 7, 7, 5, 8, 5),
  reader2 = c(8, 10, 7, 6, 6, 10, 7, 9, 8, 10, 10, 5, 6, 7, 9, 7, 7, NA, NA, NA),
  reader3 = c(7, 10, 7, 6, 6, 8, 7, 9, 8, 10, 10, 5, 6, 7, NA, NA, NA, 5, 8, 5)
)
write_files(dat = data_test, dir = tempdir())
run(dir = tempdir())

}
\seealso{
\code{\link[=load_data]{load_data()}}, \code{\link[=tally_repeats]{tally_repeats()}}, \code{\link[=write_specs_file]{write_specs_file()}}
}
\author{
Ian G. Taylor, James T. Thorson, Ian J. Stewart, Andre E. Punt
}
