% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_specs.R
\name{CreateSpecs}
\alias{CreateSpecs}
\title{Deprecated function renamed to load_specs()}
\usage{
CreateSpecs(...)
}
\arguments{
\item{...}{Any arguments associated with the deprecated function}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
CreateSpecs() has been renamed as \code{\link[=load_specs]{load_specs()}} to better clarify its purpose
}
\seealso{
\code{\link[=load_specs]{load_specs()}}
}
\author{
Ian G. Taylor
}
