% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOutputFn.R, R/TMB_AEP_fns.R
\name{PlotOutputFn}
\alias{PlotOutputFn}
\title{Plot output}
\usage{
PlotOutputFn(
  Data,
  IDataSet,
  MaxAge,
  Report,
  PlotType = "PNG",
  SaveFile = getwd(),
  subplot = 1:3,
  Nparameters = 0,
  LogLike = 0,
  ReaderNames = NULL,
  Species = NULL,
  SaveDir = "",
  Verbose = F,
  ...
)

PlotOutputFn(
  Data,
  IDataSet,
  MaxAge,
  Report,
  PlotType = "PNG",
  SaveFile = getwd(),
  subplot = 1:3,
  Nparameters = 0,
  LogLike = 0,
  ReaderNames = NULL,
  Species = NULL,
  SaveDir = "",
  Verbose = F,
  ...
)
}
\arguments{
\item{Data}{Input data matrix}

\item{MaxAge}{Maximum estimated age}

\item{PlotType}{Type of saved plots, i.e. PDF or PNG}

\item{SaveFile}{Directory for fitted model}

\item{subplot}{Vector of which plots to create.}

\item{ReaderNames}{Vector with names of each reader, defaults to
"Reader 1", "Reader 2", etc.}

\item{...}{Additional arguments passed to \code{ageing_comparison()}.}

\item{dots}{Additional arguments passed to the
\code{\link{ageing_comparison}} function.}
}
\value{
Returns AIC, AICc, and BIC for fitted model.

Returns AIC, AICc, and BIC for fitted model.
}
\description{
Plots age comparisons and results from the fitted model. Comparisons must be
conditioned on a true age that is not observed. And, in place of a true age,
the diagnostic plots generally condition on an estimated age, which is fixed
as the mode of the conditional probability at age for each otolith.

Plots age comparisons and results from the fitted Ageing Error model
}
\details{
\enumerate{
\item Error and bias by reader/lab: A panel graph is provided where each panel
shows the expected and standard deviation in age reads for that
reader/lab. This is displayed against a scatter plot of the read and
estimated ages for each otolith that was read by that reader/lab.
\item Proportion-at-age histogram: The estimated proportion at age can be
plotted as a histogram and is displayed against the observed distribution
of read ages. This is useful to determine if hte estimated proportion at
age is generally plausible, e.g., whether it has too many ages where the
estimated proportion at age approaches zero, which is unlikely in a
composite sample with moderate effective sample sizes. This plot can also
be used as a diagnostic to confirm that AIC has selected reasonable
values for the \code{MinusAge} and \code{PlusAge} parameters.
}

The function will read in \code{XXX.rep} and \code{XXX.par} files that are located in
\code{SaveFile}.
}
\references{
Punt, A.E., Smith, D.C., KrusicGolub, K., and Robertson, S. 2008.
Quantifying age-reading error for use in fisheries stock assessments,
with application to species in Australias southern and eastern scalefish
and shark fishery. Can. J. Fish. Aquat. Sci. 65: 1991-2005.

Punt, A.E., Smith, D.C., KrusicGolub, K., and Robertson, S. 2008.
Quantifying age-reading error for use in fisheries stock assessments,
with application to species in Australias southern and eastern scalefish
and shark fishery. Can. J. Fish. Aquat. Sci. 65: 1991-2005.
}
\seealso{
\itemize{
\item \code{RunFn()}
\item \code{StepwiseFn()}
}
}
\author{
James T. Thorson, Ian G. Taylor
}
