% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOutputFn.R
\name{PlotOutputFn}
\alias{PlotOutputFn}
\title{Plot output}
\usage{
PlotOutputFn(Data, MaxAge, SaveFile, PlotType = "PNG", subplot = 1:3,
  ReaderNames = NULL, ...)
}
\arguments{
\item{Data}{Input data matrix}

\item{MaxAge}{Maximum estimated age}

\item{SaveFile}{Directory for fitted model}

\item{PlotType}{Type of saved plots, i.e. PDF or PNG}

\item{subplot}{Vector of which plots to create.}

\item{ReaderNames}{Vector with names of each reader, defaults to
"Reader 1", "Reader 2", etc.}

\item{dots}{Additional arguments passed to the
\code{\link{ageing_comparison}} function.}
}
\value{
Returns AIC, AICc, and BIC for fitted model.
}
\description{
Plots age comparisons and results from the fitted Ageing Error model
}
\references{
Punt, A.E., Smith, D.C., KrusicGolub, K., and Robertson, S. 2008.
Quantifying age-reading error for use in fisheries stock assessments,
with application to species in Australias southern and eastern scalefish
and shark fishery. Can. J. Fish. Aquat. Sci. 65: 1991-2005.
}
\author{
James T. Thorson, Ian G. Taylor
}
