% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run ageing error routine}
\usage{
run(file_data, file_specs, directory)
}
\arguments{
\item{file_data}{A string specifying the file path to a data file.}

\item{file_specs}{A string specifying the file path to the specifications
file.}

\item{directory}{A string specifying a file path to a directory where you
would like to save the results.}
}
\value{
Invisibly return model output.
}
\description{
A wrapper for running a TMB model to estimate ageing error for a given data
set and specification file.
}
\author{
Kelli F. Johnson
}
