% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run ageng error routine}
\usage{
run(directory, file_data, file_specs)
}
\arguments{
\item{directory}{A string specifying a file path to a directory where you
would like to save the results.}

\item{file_data}{A string specifying the data file within 'directory'.}

\item{file_specs}{A string specifying the specifications file within 'directory'.}
}
\value{
Invisibly return model output.
}
\description{
A wrapper for running a TMB model to estimate ageing error for a given data
set and specification file.
}
\author{
Kelli F. Johnson
}
