% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepwiseFn.R
\name{StepwiseFn}
\alias{StepwiseFn}
\title{Deprecated function replaced by stepwise()}
\usage{
StepwiseFn(...)
}
\arguments{
\item{...}{Any arguments associated with the deprecated function}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
StepwiseFn() has been replaced by \code{\link[=stepwise]{stepwise()}}
}
\seealso{
\code{\link[=stepwise]{stepwise()}}
}
\author{
James T. Thorson
}
